requirements.txt
tuya-iot-py-sdk # git@github.com:tuya/tuya-iot-python-sdk.git
